Added shopping list app with list view and detail view.
