+++ ingredients/models.py
@@ -13,10 +13,3 @@ class Ingredient(models.Model):
 
     def __str__(self):
         return self.name
-
-class ShoppingList(models.Model):
-    name = models.CharField(max_length=100)
-    ingredients = models.ManyToManyField(Ingredient)
-
-    def __str__(self):
-        return self.name
