+++ ingredients/views.py
@@ -1,7 +1,8 @@
 from django.urls import reverse_lazy
 from django.http import HttpResponseRedirect
 from django.views import generic
-from .models import Ingredient, ShoppingList
+from .models import Ingredient
+from shoppingList.models import ShoppingList
 
 class Add(generic.CreateView):
     fields = ["name", "kcal_per_100g", "category"]
