+++ shoppingList/templates/shoppingList/shoppinglist_detail.html
@@ -1,0 +1,4 @@
+{{ shoppinglist.name }}
+{% for ingredient in shoppinglist.ingredients.all %} 
+<br>{{ ingredient.name }}
+{% endfor %}
