+++ shoppingList/templates/shoppingList/shoppinglist_list.html
@@ -1,0 +1,4 @@
+<h1>Shopping lists<br></h1>
+    {% for list in object_list %}
+        {{ list.name }}
+    {% endfor %}
